+++ packages/wonder-blocks-cell/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @khanacademy/wonder-blocks-cell
 
+## 4.1.22
+
+### Patch Changes
+
+- Updated dependencies [0f4f771]
+    - @khanacademy/wonder-blocks-tokens@11.1.0
+    - @khanacademy/wonder-blocks-clickable@7.1.9
+    - @khanacademy/wonder-blocks-layout@3.1.20
+    - @khanacademy/wonder-blocks-typography@4.2.4
+
 ## 4.1.21
 
 ### Patch Changes