+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.1.21
+
+### Patch Changes
+
+- Updated dependencies [0f4f771]
+    - @khanacademy/wonder-blocks-tokens@11.1.0
+    - @khanacademy/wonder-blocks-typography@4.2.4
+
 ## 5.1.20
 
 ### Patch Changes